Fix: usBackGround is now UINT16 (was UINT8) to accommodate more than 255 backgrounds.
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@6513 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Laptop/CharProfile.cpp
@@ -79,7 +79,7 @@ INT32 iPersonality = 0;
 INT32 iAttitude = 0;
 
 // Flugente: background
-UINT8 usBackground = 0;
+UINT16 usBackground = 0;
 
 // Flugente: sexism, racism etc.
 INT8  bRace = 0;
